✅🚚 Move argument sending tests to their own file
